Moved some tests utilities from 'test_zipfile/...' to the general 'support' folder
