+++ Lib/test/test_zipfile/_path/_itertools.py
@@ -29,20 +29,6 @@ class Counter:
         return result
 
 
-# from more_itertools v8.13.0
-def always_iterable(obj, base_type=(str, bytes)):
-    if obj is None:
-        return iter(())
-
-    if (base_type is not None) and isinstance(obj, base_type):
-        return iter((obj,))
-
-    try:
-        return iter(obj)
-    except TypeError:
-        return iter((obj,))
-
-
 # from more_itertools v9.0.0
 def consume(iterator, n=None):
     """Advance *iterable* by *n* steps. If *n* is ``None``, consume it
